\name{eolts-package}
\alias{eolts-package}
\alias{eolts}
\docType{package}
\title{
Support for reading, writing, and analysis and plotting of time series and spectra.
}
\description{
This package provides R functions to read, analyse, plot and write time series data, and to generate and plot spectra from time series. Support includes reading time series from NetCDF files, and writing time series to a NetCDF RPC server. Developed at the National Center for Atmospheric Research, Earth Observing Laboratory (NCAR EOL), sponsored by the National Science Foundation.
}
\details{
\tabular{ll}{
Package: \tab eolts\cr
Version: \tab 0.0-1\cr
Date: \tab 2013-09-13\cr
Authors@R: \tab person("Gordon","Maclean",role=c("aut","cre"),
            email="maclean@ucar.edu")\cr
Depends: \tab R (>= 3.0), methods\cr
Imports: \tab splusTimeDate,splusTimeSeries\cr
Suggests: \tab RUnit\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.eol.ucar.edu/\cr
}
}
\author{
Gordon Maclean

Maintainer: Gordon Maclean <maclean@ucar.edu>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{

}
